Adding waypoint logic into a single file.
